[PixelOverlay] Modify Pixel Map to support more than 64K channels

+++ src/memorymap.c
@@ -46,7 +46,7 @@ char         *chanDataMap;
 int           chanDataMapFD = -1;
 char         *ctrlMap;
 int           ctrlFD = -1;
-short        *pixelMap;
+long long    *pixelMap;
 int           pixelFD = -1;
 
 FPPChannelMemoryMapControlHeader *ctrlHeader = NULL;
@@ -61,10 +61,10 @@ int LoadChannelMemoryMapData(void);
 int InitializeChannelDataMemoryMap(void) {
 	int   i  = 0;
 	char  ch = '\0';
-	short pixelLocation = 0;
+	long long pixelLocation = 0;
 	char  tmpData[FPPD_MAX_CHANNELS];
 	char  tmpCtrlData[FPPCHANNELMEMORYMAPSIZE];
-	short tmpPixelMapData[FPPD_MAX_CHANNELS];
+	long long tmpPixelMapData[FPPD_MAX_CHANNELS];
 
 	LogDebug(VB_CHANNELOUT, "InitializeChannelDataMemoryMap()\n");
 
@@ -150,7 +150,7 @@ int InitializeChannelDataMemoryMap(void) {
 		return -1;
 	}
 
-	pixelMap = (short *)mmap(0, FPPD_MAX_CHANNELS * sizeof(short), PROT_READ|PROT_WRITE, MAP_SHARED, pixelFD, 0);
+	pixelMap = (long long *)mmap(0, FPPD_MAX_CHANNELS * sizeof(long long), PROT_READ|PROT_WRITE, MAP_SHARED, pixelFD, 0);
 
 	if (!pixelMap) {
 		LogErr(VB_CHANNELOUT, "Error mapping %s memory map file: %s\n",
